ADD README.md and .gitignore
First commit for the repository.
A draft for README.md was added as well as a standard .gitignore
file.
The TODO document was created.
My personal utility library (libft) was added as a submodule.

From now on every commit will be on the development branch or its
succesors in a per feature basis. Only very stable and clean versions
will be merged back to the main branch.