ADD Receiving & interpreting note on/off midi msg
This dirty version is able to locate the device file for MIDI
devices (for ALSA driver) though some hints were included to
support older drivers (e.g. OSS). It can open the device
file and listen to the MIDI messages sent by the keyboard.
So far it supports two types of messages: channel/voice messages:
note on (key is pressed) and note off (key is released).
The program is able to parse these two types of messages and
to decode the key id into a latin note naming style.

This is a draft to proof the concept. The next steps will
probably be to clean it up, re-factor it and probably to advance
into a basic trainer so I can begin to use it to practice my
piano.