Refactorized, multithread, working note training
The midi interpreter now runs as an independent thread
which made possible to implement the first excercise: play the note.

A random note generator was added.

Now the SIGINT is catched for a controlled exit using an
exit pending flag.

The device file is now opened with O_NONBLOCK to allow a faster
termination of the MIDI listener thread.

No memory leaks detected. A much more cleaner code.