ADD terminal capabilities features
User interface experiments just to play with the following functions
(that I need in another project):
        From unistd.h or C standard library"
                isatty
                ttyname
                tcsetattr
                tcgetattr
                ttyslot
                ioctl
        From termcap library:
                tgetent
                tgetflag
                tgetnum
                tgetstr
                tgoto
                tputs

Controlled initialization and termination of the user's interface.
The ECHO capabilities of the terminal is disabled on init.
All three standard terminal-related file descriptors are checked
and the program only launches if all of them are a tty.
The original configuration of the terminal is restored on termination.
The screen is cleared in the beginning and before terminating the
program.

First doxyfile aproximation (the header files headers and
little more).

Re-organization of the header files to avoid forward declarations
for the t_pt type (and probably more to come in the future).