aFIX Refactorization, better encapsulation
New stuctures were created for midi and the guess_note game to store
members previously located in the main pt struct. Creator and
destroyer functions were written for this new structs and the ones
for pt were modified accordingly.

Slowly improving the old midi, but not yet re-designed. It almost
works as I wanted, but it's not written as I want it.

Lot's of known leaks (I just wanted to play a little).