CLEANING tyding things up before my trip
Still re-writing the guess note game under the new paradigm.
So far it shows some (bad) random note and the last note
received from midi. It does so in real-time.

May be this is a good time for going back to termcap to see
if theres an alterrative for the re-renderization of the
full screen.

RESUME at pt_gm_get_rand_note.c and pt_gm_guess_note.c
in the context of using the configuration settings of
the game to: (1) Randomly pick a note within the parameters;
(2) Render it according to the settings. Implement the states
of the game.